f1_weather_agent/requirements.txt
pydantic
langgraph
langgraph-cli[inmem]
langchain-core
langchain-openai
fastmcp
a2a-sdk
uvicorn
